+++ tsconfig.json
@@ -1,14 +1,15 @@
 {
 	"compilerOptions": {
 		"module": "commonjs",
-		"target": "es6",
+    "target": "es6",
+    "allowJs": true,
 		"outDir": "out",
 		"lib": [
 			"es6"
     ],
     "moduleResolution": "node",
 		"sourceMap": true,
-		"rootDir": ".",
+		"rootDir": "src",
 		"strict": true,   /* enable all strict type-checking options */
 		/* Additional Checks */
 		"noImplicitReturns": true, /* Report error when not all code paths in function return a value. */
@@ -18,5 +19,5 @@
 		"node_modules",
     ".vscode-test",
     "webpack.config.ts"
-	]
+  ]
 }
